pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neuroclaw"
version = "0.1.0"
description = "Visual Voxel Micro-Engine — Stage 1 multimodal TRIBE v2 extraction"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "NeuroClaw" }]
dependencies = [
    "torch>=2.1",
    "transformers>=4.36",
    "safetensors>=0.4",
    "zstandard>=0.22",
    "nilearn>=0.10",
    "nibabel>=5.0",
    "numpy>=1.23",
    "pandas>=2.0",
    "opencv-python-headless>=4.8",
    "ffmpeg-python>=0.2",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "typer>=0.9",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
# Same-interpreter ASR (bypasses tribev2 uvx/whisperx); recommended on macOS / NumPy 2 stacks.
local-asr = ["faster-whisper>=1.0.0"]
ocr = ["pytesseract>=0.3.10"]
tribev2 = ["tribev2>=0.3", "faster-whisper>=1.0.0", "pytesseract>=0.3.10"]
mlx = ["mlx>=0.14"]
dev = [
    "pytest>=7.4",
    "ruff>=0.1",
    "mypy>=1.7",
    "faster-whisper>=1.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
neuroclaw-extract = "neuroclaw.cli:main"

[tool.hatch.build.targets.sdist]
include = ["src/neuroclaw", "README.md", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
packages = ["neuroclaw"]
